requirements.txt
anthropic==0.40.0
fastapi==0.115.6
uvicorn[standard]==0.32.1
pypdf==5.1.0
python-multipart==0.0.20
pydantic==2.10.3
mangum==0.19.0
python-dotenv==1.0.1
